adding changes
adding changes

+++ Pac_Man/player.cpp
@@ -4,21 +4,19 @@ using namespace sf;
 using namespace std;
 
 void Player::Update(double dt) {
-
-    // Move in four directions based on keys
-    if (Keyboard::isKeyPressed(Keyboard::Up)) {
-        _shape->move(0.0f, -_speed * dt); // Move up
-    }
-    if (Keyboard::isKeyPressed(Keyboard::Down)) {
-        _shape->move(0.0f, _speed * dt); // Move down
-    }
-    if (Keyboard::isKeyPressed(Keyboard::Left)) {
-        _shape->move(-_speed * dt, 0.0f); // Move left
-    }
-    if (Keyboard::isKeyPressed(Keyboard::Right)) {
-        _shape->move(_speed * dt, 0.0f); // Move right
-    }
-
+   // Move in four directions based on keys
+   if (Keyboard::isKeyPressed(Keyboard::Up)) {
+       _shape->move(0.0f, -_speed * static_cast<float>(dt)); // Move up
+   }
+   if (Keyboard::isKeyPressed(Keyboard::Down)) {
+       _shape->move(0.0f, _speed * static_cast<float>(dt)); // Move down
+   }
+   if (Keyboard::isKeyPressed(Keyboard::Left)) {
+       _shape->move(-_speed * static_cast<float>(dt), 0.0f); // Move left
+   }
+   if (Keyboard::isKeyPressed(Keyboard::Right)) {
+       _shape->move(_speed * static_cast<float>(dt), 0.0f); // Move right
+   }
 
 
     Entity::Update(dt);
